refactor(scripts): refactor on script files and remove unused files
