--- src/scripts/index.ts
@@ -1,1 +1,0 @@
-import main as Borrow from "./aaveBorrow";
