--- src/scripts/server.ts
@@ -1,5 +1,0 @@
-const logger = require("pino")();
-
-(async () => {
-  logger.info("Running the server...");
-})();
